db_setup/requirements.txt
pydantic==2.12.5
SQLAlchemy==2.0.49
sqlmodel==0.0.38
python-dotenv==1.2.2
psycopg2-binary==2.9.11